\name{exportCytoGML}
\alias{exportCytoGML}
\title{
Write out an igraph graph to Cytoscape readable GML file.
}
\description{
Takes an igraph object and writes it to a GML file that is readable by Cytoscape.
}
\usage{
exportCytoGML(graph, filename)
}
\arguments{
  \item{graph}{
	\code{graph} igraph graph
}
  \item{filename}{
	\code{filename} output filename
}
}
\author{
Markus Schroeder <mschroed@jimmy.harvard.edu>

Benjamin Haibe-Kains <bhaibeka@jimmy.harvard.edu>
}
\examples{
## todo
}
\keyword{ igraph }
\keyword{ graph }
\keyword{ Cytoscape }
\keyword{ network }
