\name{igraph-class}
\docType{class}
\alias{igraph-class}

\title{Class \code{"igraph"}}
\description{
This is an igraph object.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"}}
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
  See \code{\link[igraph:igraph]{igraph}} for functions.
}
\references{
  See \code{\link[igraph:igraph]{igraph}}
}
\keyword{classes}
