\name{RamiGO}
\alias{RamiGO}
\docType{package}
\title{
amiGO R interface
}
\description{
R interface sending requests to amiGO and retrieving DAG GO trees.
}
\details{
\tabular{ll}{
Package: \tab RamiGO\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-05-27\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Markus Schroeder

Maintainer: Markus Schroeder <mschroed@jimmy.harvard.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
