\name{RamiGO}
\alias{RamiGO}
\docType{package}
\title{
AmiGO visualize R interface
}
\description{
R interface sending requests to AmiGO visualize, retrieving DAG GO trees, parsing GraphViz DOT format files and exporting GML files for Cytoscape.
}
\details{
\tabular{ll}{
Package: \tab RamiGO\cr
Type: \tab Package\cr
Version: \tab \Sexpr{packageDescription("RamiGO")$Version}\cr
Date: \tab \Sexpr{packageDescription("RamiGO")$Date}\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
\bold{Markus Schroeder}

- Biostatistics and Computational Biology, Dana-Farber Cancer Institute, Harvard School of Public Health, Boston, MA, USA

\url{http://compbio.dfci.harvard.edu/}

- Computational Genomics, Center for Biotechnology (CeBiTec), Bielefeld University, Germany

\url{http://www.cebitec.uni-bielefeld.de/cebitec/computational-genomics/home.html}

\bold{Maintainer}:

Markus Schroeder

\email{mschroed@jimmy.harvard.edu}
}
\references{
Gene Ontology: tool for the unification of biology. The Gene Ontology Consortium (2000) Nature Genet. 25: 25-29

http://amigo.geneontology.org/cgi-bin/amigo/go.cgi
}
